V1.67 Server框架提供新功能,用户数据于NetHost关联的相关接口
V1.67
1.服务器框架改动
1.1增加表示连接数据的基础类HostData
用户直接从HostData派生自己的数据类型即可
通过NetHost::SetData()接口将数据与连接关联
具体使用方法参考头文件
1.2NetHost的IO接口优化
数据长度，从原来的unsigned short扩大到unsigned int

2.优化Logger日志类
2.1增加3个接口
SetMaxLogSize()日志最多保存这个大小的内容
SetMaxExistDay()日志最多存在这么多天
SetLogName()日至名字，会产生一个日志名的文件夹
2.1写各种类型日志接口，统一为2个
Info()输出日志
StreamInfo()输出流

3.扩展ConfigFile配置文件类
增加对Section的支持
使用范例
配置文件内容
[ser config]
ip = 192.168.0.1
port = 8888
[/ser config]

ConfigFile cfg( "./test.cfg" );
string ip = cfg["ser config"]["ip"];
cfg["ser config"]["ip"] = "127.0.0.1";
cfg["ser config"]["port"] = 8080;//可以直接使用char*、string、int、float等进行赋值
cfg.save();

4.ThreadPool线程池增加1个接口，GetTaskCount取得当前任务数

5.Socket类
5.1优化Connect()接口，可以直接传递域名进行连接
5.2增加static接口HostName2IP()，将域名转换成IP

6.mapi.h增加几个公用api
6.1mdk_assert代替assert()
6.2GetFileSize
6.3GetCUPNumber
6.4CurThreadId
6.5mdk_Date

7.优化一些造成编译警告的代码

+++ include/frame/netserver/STNetConnect.h
@@ -54,8 +54,8 @@ public:
 	uint32 GetLength();//取得数据长度
 	//从接收缓冲中读数据，数据不够，直接返回false，无阻塞模式
 	//bClearCache为false，读出数据不会从接收缓冲删除，下次还是从相同位置读取
-	bool ReadData(unsigned char* pMsg, unsigned short uLength, bool bClearCache = true );
-	bool SendData( const unsigned char* pMsg, unsigned short uLength );
+	bool ReadData(unsigned char* pMsg, unsigned int uLength, bool bClearCache = true );
+	bool SendData( const unsigned char* pMsg, unsigned int uLength );
 	bool SendStart();//开始发送流程
 	void SendEnd();//结束发送流程
 	void Close();//关闭连接
