+++ source/mdk/IOBufferBlock.cpp
@@ -40,11 +40,11 @@ void IOBufferBlock::WriteFinished(unsigned short uLength)
 {
 	m_uLength += uLength;
 }
- 
+
 //冲缓冲块读取uLength长度的数据
-unsigned short IOBufferBlock::ReadData( unsigned char *data, unsigned short uLength, bool bDel )
+unsigned int IOBufferBlock::ReadData( unsigned char *data, unsigned int uLength, bool bDel )
 {
-	unsigned short uReadSize;
+	unsigned int uReadSize;
 	if ( uLength > m_uLength - m_uRecvPos ) uReadSize = m_uLength - m_uRecvPos;
 	else uReadSize = uLength;
 	memcpy( data, &m_buffer[m_uRecvPos], uReadSize );
